rename LocationShift to ShiftUtils and some method alignments
